requirements.txt
PySide==1.2.2
argparse==1.2.1
pyserial==2.7
wsgiref==0.1.2
